Find and print the first 98 Finoncci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	long double c, s, f1, f2;
+	long int c, s, f1, f2;
 
 	f1 = 1;
 	f2 = 2;
